requirements.txt
flask==3.0.0
pymongo==4.6.1
python-dotenv==1.0.0
faster-whisper
yt-dlp
pyannote.audio==3.1.1
transformers
torch
torchaudio
accelerate
vaderSentiment

sentence-transformers
bertopic
pandas
fpdf
python-docx
marshmallow
requests
scikit-learn
google-generativeai
